refactor(search): nest Search into folders by sub-namespace
Mirror the Spotify/YouTube/Apple/Reddit technical-layer layout for RedditPodcastPoster.Search:

- Configuration/: SearchIndexConfig

- Models/: EpisodeSearchRecord, IndexerState, IndexerStateWrapper, SearchDocument

- Factories/: ISearchClientFactory+impl, ISearchIndexClientFactory+impl, ISearchIndexerClientFactory+impl

- Services/: ISearchIndexerService (renamed file from ISearchIndexerServices.cs), SearchIndexerService

- Formatting/: Constants, DescriptionTruncator

- Extensions/: ServiceCollectionExtensions (unchanged location)

Public type names and the AddSearch() DI extension are unchanged. Updated all consumer usings across Console-Apps, Cloud/Indexer, Cloud/Api and RedditPodcastPoster.EntitySearchIndexer, including alias fixes for the IndexerState naming collision with Api.Dtos.IndexerState.
