+++ Class-Libraries/RedditPodcastPoster.Search/Factories/ISearchIndexClientFactory.cs
@@ -1,6 +1,6 @@
 ﻿using Azure.Search.Documents.Indexes;
 
-namespace RedditPodcastPoster.Search;
+namespace RedditPodcastPoster.Search.Factories;
 
 public interface ISearchIndexClientFactory
 {
