+++ Class-Libraries/RedditPodcastPoster.Search/Factories/ISearchIndexerClientFactory.cs
@@ -1,6 +1,6 @@
 ﻿using Azure.Search.Documents.Indexes;
 
-namespace RedditPodcastPoster.Search;
+namespace RedditPodcastPoster.Search.Factories;
 
 public interface ISearchIndexerClientFactory
 {
