+++ Class-Libraries/RedditPodcastPoster.Search/Factories/SearchClientFactory.cs
@@ -2,8 +2,9 @@
 using Azure.Search.Documents;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
+using RedditPodcastPoster.Search.Configuration;
 
-namespace RedditPodcastPoster.Search;
+namespace RedditPodcastPoster.Search.Factories;
 
 public class SearchClientFactory(
     IOptions<SearchIndexConfig> searchIndexConfig,
