+++ Class-Libraries/RedditPodcastPoster.Search/Factories/SearchIndexClientFactory.cs
@@ -2,8 +2,9 @@
 using Azure.Search.Documents.Indexes;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
+using RedditPodcastPoster.Search.Configuration;
 
-namespace RedditPodcastPoster.Search;
+namespace RedditPodcastPoster.Search.Factories;
 
 public class SearchIndexClientFactory(
     IOptions<SearchIndexConfig> searchIndexConfig,
