--- Class-Libraries/RedditPodcastPoster.Search/ISearchIndexerServices.cs
@@ -1,6 +1,0 @@
-﻿namespace RedditPodcastPoster.Search;
-
-public interface ISearchIndexerService
-{
-    Task<IndexerStateWrapper> RunIndexer();
-}
