+++ Class-Libraries/RedditPodcastPoster.Search/Models/IndexerStateWrapper.cs
@@ -1,4 +1,4 @@
-﻿namespace RedditPodcastPoster.Search;
+﻿namespace RedditPodcastPoster.Search.Models;
 
 public record IndexerStateWrapper(
     IndexerState IndexerState,
