+++ Class-Libraries/RedditPodcastPoster.Search/Services/ISearchIndexerService.cs
@@ -1,0 +1,8 @@
+﻿using RedditPodcastPoster.Search.Models;
+
+namespace RedditPodcastPoster.Search.Services;
+
+public interface ISearchIndexerService
+{
+    Task<IndexerStateWrapper> RunIndexer();
+}
