+++ Console-Apps/CreateSearchIndex/CreateSearchIndexProcessor.cs
@@ -12,7 +12,9 @@ using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using RedditPodcastPoster.Models.Extensions;
 using RedditPodcastPoster.Persistence;
-using RedditPodcastPoster.Search;
+using RedditPodcastPoster.Search.Formatting;
+using RedditPodcastPoster.Search.Models;
+using RedditPodcastPoster.Search.Services;
 
 namespace CreateSearchIndex;
